Rebuild commit history
Rebuilt the commit history to follow a better git
workflow. Originally all branches worked off
master with no clear order of what was ready
for release and what was being worked on.

This branch as it stands now is operable and
works, to the best of my knowledge, as expected.

This is an effort to better streamline the pkg.