Remove xml2 from Imports
The most recent successful travis test, https://goo.gl/XMhlnK, did
not have xml2 package in the DESCRIPTION file. This test passed
though because in the .travis.yml file the warnings_are_errors
param is set to false.

When running a R CMD check a warning is generated:

```
checking dependencies in R code ... WARNING
'::' or ':::' import not declared from: ‘xml2’
```

So, first I'll see if this tests passes. Then, if so, will address
fixing the check issue.