Rewrote vignette for minimal reproducible code
This is a base vignette centered around the code issue. When we look
at the /inst/docs/getting-started.html we can see all code functions
as expected. There are no errors. The various methods of assining
the link parameter in get_storm_data all work.

See https://goo.gl/SxUvtQ for results of this push in travis.