Modify date/time formatting in fstadv_forecasts
Since forecast points are only given a date and time (in tz Z) original
script would increment month by 1 if day was less than the observation
day and year by 1 if month was less than the observation month.

Tropical Storm Zeta (AL302005) is the first storm in this archive to
cross years. This exposed the month being incremented to a value of 13
which would fail on the lubridate::ymd_hm() call. Thus, had to
modify this portion of script to subtract 12 from mo if mo > 12.

All tests pass.