Script to loop through basins, years, prods; save
This script will loop through all available combinations of options
for storm data and save that data (if it exists) to a data directory.

This is not a data validation tool.

It can be modified by Variables to work a specific basin, year or
product or left as-is to run on all products, all basins, all years.
The latter is very time-consuming so grab a coffee or multi-task.