Remove progress bar from get_url_contents
The way the previous setup worked, the progress bar was representing
extracting the URLs from the NHC archives. However, this did not
represent the parsing of the data that was taking place. This would be
misleading to users (and frustrating as I found out).

Instead, the data gathering functions have been split up a bit to help
the progress bar environment track the entire process of data
gathering.