Add func `get_storm_list` (#114)
Add function `get_storm_list` which will retrieve what seems to be a
list of all known cyclones, regardless of basin, from the NHC FTP
website.

Added documentation explaining the dataframe variables.

Added tests for structure of the returned object, as well as expected
names for the returned object. Cannot test row count as it is
evergreen.

Bumped version to 0.2.0-6.9000