Update tests for changes in commit 269e71a (#119)
Test for function `get_nhc_link` was failing; I did not update it as
I should.

Additionally, the gis function tests which verify the URLs have not
changed to random data sets still failed. This is because the NHC
is not using the https links for these datasets. Rewrote these
tests to use a non-https URL.