Remove time restrictions in `get_storms`
There is a request from the NHC to limit web hits to 80 per 10 seconds.
By default, the `get_storms` function would split URLs into groups of
four and send those links to `get_storms_link` with a delay of half
a second. This was stupid on my part as, for this call, there will not
be a request of 80 links until the year 2078.

So, I fixed it.

Additionally, did some code cleaning.