Return text if not within html; else extract node
The previous version did not handle advisories that were not embedded
in HTML. Wrapped `xml2::read_html` within a `purrr::safely` func. If
the error object is not null, then we're working on a non-HTML
product. Otherwise, we are on a HTML page and need to extract the
text from within the HTML.