Correct invalid time capture in header, forecasts
For forecast advisories, the absence of AM|PM caused the calculation
to assume all times were AM. Pattern has been modified to make the
capture optional.

Additionally, for forecast times, incorrect math was applied that was
generating bad times for the forecast periods. The math has been
removed.

Instead, I've modified the code which already extracted the day (Date),
Hour, and Minute and ran case_when functions. If the Date value of the
forecast period is less than the lubridate::day() value of AdvDate,
then the forecast period goes into the proceeding month; therefore,
the month value is extracted from `AdvDate` and incremented by one;
otherwise, `FcstMonth` is the month value of `AdvDate`.

Similar calculation was applied to ensure the proper year was
calculated. While rare a storm will be active from one year to the
next, it has happened. This calculation should ensure that is
accounted.